+++ Pandora/Paradigm/Inventory/Environment.hs
@@ -4,13 +4,15 @@ module Pandora.Paradigm.Inventory.Environment (Environment (..), Configured, env
 
 import Pandora.Pattern.Category (identity, (.), ($))
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>)), Covariant_ ((-<$>-)))
+import Pandora.Pattern.Functor.Contravariant (Contravariant_ ((->$<-)))
 import Pandora.Pattern.Functor.Pointable (Pointable (point))
 import Pandora.Pattern.Functor.Applicative (Applicative ((<*>)))
 import Pandora.Pattern.Functor.Distributive (Distributive ((-<<)))
 import Pandora.Pattern.Functor.Bindable (Bindable ((>>=)))
 import Pandora.Pattern.Functor.Monad (Monad)
-import Pandora.Pattern.Functor.Divariant (Divariant ((>->)))
+import Pandora.Pattern.Functor.Divariant (Divariant_ ((->->-)))
 import Pandora.Paradigm.Primary.Functor.Function ((!.), (%))
+import Pandora.Paradigm.Primary.Transformer.Flip (Flip (Flip))
 import Pandora.Paradigm.Controlflow.Effect.Interpreted (Schematic, Interpreted (Primary, run, unite))
 import Pandora.Paradigm.Controlflow.Effect.Transformer.Monadic (Monadic (wrap), (:>) (TM))
 import Pandora.Paradigm.Controlflow.Effect.Adaptable (Adaptable (adapt))
@@ -23,6 +25,9 @@ instance Covariant (Environment e) where
 
 instance Covariant_ (Environment e) (->) (->) where
 	f -<$>- Environment x = Environment $ f . x
+
+instance Contravariant_ (Flip Environment a) (->) (->) where
+	f ->$<- Flip (Environment g) = Flip . Environment $ g . f
 
 instance Pointable (Environment e) (->) where
 	point x = Environment (x !.)
@@ -38,8 +43,8 @@ instance Bindable (Environment e) where
 
 instance Monad (Environment e) where
 
-instance Divariant Environment where
-	(>->) ab cd bc = Environment $ ab >-> cd $ run bc
+instance Divariant_ Environment (->) (->) (->) where
+	(->->-) ab cd bc = Environment $ ab ->->- cd $ run bc
 
 instance Interpreted (Environment e) where
 	type Primary (Environment e) a = (->) e a