+++ Pandora/Pattern/Functor/Distributive.hs
@@ -1,6 +1,5 @@
 module Pandora.Pattern.Functor.Distributive where
 
-import Pandora.Core.Functor (type (:.), type (:=))
 import Pandora.Pattern.Functor.Covariant (Covariant_)
 
 {- |