+++ CHANGELOG.md
@@ -500,3 +500,4 @@
 * Rename experimental `Bivariant_` to `Bivariant`
 * Remove `Distributive` typeclass
 * Rename experimental `Distributive_` to `Distributive`
+* Remove `Divariant` typeclass